Update CrashHandler.c
Removed mutex unlock and destroy - may cause a small memory leak, but this is a crash, and so that will already be happening (memory leaks are cleared up when the program is terminated). The alternative is risking the code unlocking prior to exiting the program, causing an unnecessary unsafe crash.